\docType{data}
\name{benichou}
\alias{benichou}
\title{Data in table 1 in Benichou (1991) from study of esophageal
  cancer (Tuyns, et al.)}
\format{A data frame with 975 subjects and 8 variables.
   \describe{
       \item{cases}{ Esophageal cancer indicator (1=case, 0=control). }
       \item{age}{ Subject age in years. }
       \item{smoke}{ Quantity of tobacco smoked per day. (0-9g/day,
	10-29g/day, or 30+g/day) }
       \item{alcohol}{ Quantity of alcohol consumed per day
	 (0-39g/day, 
	 120+g/day, 40-79g/day, or 80-119g/day) }
       \item{alcohol80}{ Quantity of alcohol consumed per day
	 (0-79g or 80+g) }
       \item{smoke.alc}{ Is set to 'Exposed'
	 if subject was consumed more than 40+g of alcohol and
	 10+g of tobacco per day.}
       \item{alcohol40}{ Quantity of alcohol consumed per day
	 (0-39g or 40+g) }
       \item{smoke10}{ Quantity of tobacco consumed per day
	 (0-9g or 10+g) }
   }}
\description{
  Data in table 1 in Benichou (1991) from study of
  esophageal cancer (Tuyns, et al.)
}
\references{
  Benichou, J. (1991) Methods of adjustment for estimating
  the attributable risk in case-control studies: a review
  Statistics in Medicine 10:1753-1773.  Tuyns AJ, Pequignot
  G, Jensen OM. Esophageal cancer in Ille-et-Vilaine in
  relation to levels of alcohol and tobacco consumption.
  Risks are multiplying. Bull Cancer. 1977;64(1):45-60.
}
\keyword{data}
\keyword{datasets}

