\docType{data}
\name{chapter.dat}
\alias{chapter.dat}
\title{The data from Whisnant, et al 1996 concerning cerebral
  infarction, blood pressure, diabetes.}
\format{A data frame with 2644 subjects and 5 variables.
  \itemize{
    \item{cases}{ Stroke indicator. (1=Stroke, 0=Normal)}
    \item{hbp}{ High blood pressure indicator. (1=High Blood Pressure, 0=Normal ) }
    \item{ihd}{ Ischemic heart disease indicator. (1=Heart Disease, 0=Normal) }
    \item{dm}{ Diabetes indicator. (1=Diabetic, 0= Normal) }
    \item{match.id}{ Matched case/controls have same ID. }
}}
\description{
  The data from Whisnant, et al 1996 concerning cerebral
  infarction, blood pressure, diabetes.
}
\references{
  Whisnant, J.P., Wiebers, D.O., OFallon, W.M., Sicks,
  J.D., Frye, R.L., (1996) A population-based model of risk
  factors for ischemic stroke: Rochester, Minnesota
  Neurology 47:1420-1428.
}
\keyword{data}
\keyword{datasets}

