\name{stroke.dat}
\alias{stroke.dat}
\docType{data}
\title{Data from Davis PH, et al 1987 Risk factors for ischemic stroke: a prospective study in Rochester, Minnesota. }
\description{
Data from Davis PH, et al 1987 Risk factors for ischemic stroke: a prospective study in Rochester, Minnesota. 
}
\usage{data(stroke.dat)}
\format{
  A data frame with 3388 observations on the following 4 variables.
  \describe{
    \item{cases}{ Stroke indicator. (1=Stroke, 0=Normal) }
    \item{age}{ Years }
    \item{smoke}{ Current, Former, Never, Uncertain, Unknown, or <NA> }
    \item{diastolic}{ Diastolic blood pressure }
  }
}

\references{
Davis, P.H., Dambrosia, J.M., Schoenberg, B.S., et al, (1987) Risk factors for ischemic stroke: a
prospective study in Rochester, Minnesota Annals of Neurology 22:319-327
}
\keyword{data}
\keyword{datasets}
