\name{summary.attribrisk}
\alias{summary.attribrisk}
\title{Summarizes the attribrisk object.}
\usage{
  \method{summary}{attribrisk}(object, ...)
}
\arguments{
  \item{object}{A attribrisk output object of class
  "attribrisk".}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The attribrisk object invisible flag set to prevent
  printing.
}
\description{
  This is a method for the function summary for object of
  the class "attribrisk".
}
\seealso{
  \code{\link{attribrisk}}
}

