\name{whisnant}
\alias{whisnant}
\docType{data}
\title{ Summary data from Whisnant, et al 1996 concerning cerebral infarction, blood pressure, diabetes }
\description{ Summary data from Whisnant, et al 1996 concerning cerebral infarction, blood pressure, diabetes }
\usage{data(whisnant)}
\format{
  A data frame with 4 observations on the following 3 variables.
  \describe{
    \item{\code{infarct}}{ Stroke indicator. (1=Stroke, 0=Normal) }
    \item{\code{hbp}}{High blood pressure indicator. (1=High Blood Pressure, 0=Normal ) }
    \item{\code{y}}{ The number of patients in the observed categories }
  }
}
\keyword{datasets}
